Hierarchy updates
Refactor cross-compilation logic
