+++ core/src/main/scala/chisel3/experimental/hierarchy/core/Hierarchy.scala
@@ -1,30 +1,55 @@
+// SPDX-License-Identifier: Apache-2.0
+
 package chisel3.experimental.hierarchy.core
 
 import chisel3._
+
 import scala.collection.mutable.{HashMap, HashSet}
 import chisel3.experimental.BaseModule
 import _root_.firrtl.annotations.IsModule
-import scala.reflect.runtime.universe.TypeTag
 
-trait Hierarchy[+A] extends HierarchyImpl[A] {
+import scala.annotation.implicitNotFound
 
-  /** Determine whether underlying proto is of type provided.
+/** Super-trait for Instance and Definition
+  *
+  * Enables writing functions which are Instance/Definition agnostic
+  */
+sealed trait Hierarchy[+A] extends HierarchyIsA[A] {
+
+  /** Updated by calls to [[_lookup]], to avoid recloning returned Data's */
+  private[chisel3] val cache = HashMap[Data, Data]()
+  private[chisel3] def getInnerDataContext: Option[BaseModule]
+
+  /** Used by Chisel's internal macros. DO NOT USE in your normal Chisel code!!!
+    * Instead, mark the field you are accessing with [[public]]
     *
-    * @note IMPORTANT: this function requires summoning a TypeTag[B], which will fail if B is an inner class.
-    * @note IMPORTANT: this function IGNORES type parameters, akin to normal type erasure.
-    * @note IMPORTANT: this function relies on Java reflection for underlying proto, but Scala reflection for provided type
+    * Given a selector function (that) which selects a member from the original, return the
+    *   corresponding member from the hierarchy.
     *
-    * E.g. isA[List[Int]] will return true, even if underlying proto is of type List[String]
-    * @return Whether underlying proto is of provided type (with caveats outlined above)
+    * Our @instantiable and @public macros generate the calls to this apply method
+    *
+    * By calling this function, we summon the proper Lookupable typeclass from our implicit scope.
+    *
+    * @param that a user-specified lookup function
+    * @param lookup typeclass which contains the correct lookup function, based on the types of A and B
+    * @param macroGenerated a value created in the macro, to make it harder for users to use this API
     */
-  def isA[B: TypeTag]: Boolean = {
-    val tptag = implicitly[TypeTag[B]]
-    // drop any type information for the comparison, because the proto will not have that information.
-    val name = tptag.tpe.toString.takeWhile(_ != '[')
-    inBaseClasses(name)
-  }
+  def _lookup[B, C](
+    that: A => B
+  )(
+    implicit lookup: Lookupable[B],
+    macroGenerated:  chisel3.internal.MacroGenerated
+  ): lookup.C
 
+  /** @return Return the underlying Definition[A] of this Hierarchy[A] */
+  def toDefinition: Definition[A]
+
+  /** @return Convert this Hierarchy[A] as a top-level Instance[A] */
+  def toInstance: Instance[A]
 }
+
+// Used to effectively seal Hierarchy, without requiring Definition and Instance to be in this file.
+private[chisel3] trait SealedHierarchy[+A] extends Hierarchy[A]
 
 object Hierarchy {
   implicit class HierarchyBaseModuleExtensions[T <: BaseModule](i: Hierarchy[T]) {
